// src/generated-types/store-attribute.ts
export interface Id {
  id: string;
}

/** Message representing a status response */
export interface StatusResponse {
  success: boolean;
  message: string;
}

/** Request to get attributes by category ID */
export interface GetAttributesByCategoryIdRequest {
  categoryId: string;
}

/** Message representing a list of attributes */
export interface AttributeList {
  data: AttributeWithTranslations[];
}

/** Message representing an attribute */
export interface AttributeResponse {
  id: string;
  categoryId: string;
  slug: string;
  sortOrder: number;
}

/** Message representing an attribute with translations */
export interface AttributeWithTranslations {
  id: string;
  categoryId: string;
  slug: string;
  sortOrder: number;
  translations: AttributeTranslationResponse[];
}

/** Message representing an attribute translation */
export interface AttributeTranslationResponse {
  id: string;
  language: Language;
  name: string;
}

/** Message for creating a new attribute */
export interface CreateAttributeRequest {
  categoryId: string;
  slug: string;
}

/** Message for updating an existing attribute */
export interface UpdateAttributeRequest {
  id: string;
  slug?: string | null | undefined;
}

/** Message for changing the position of an attribute */
export interface ChangeAttributePositionRequest {
  id: string;
  sortOrder: number;
}

/** Message for creating or updating an attribute translation */
export interface AttributeTranslationRequest {
  attributeId: string;
  language: Language;
  name: string;
}

export const STORE_ATTRIBUTE_V1_PACKAGE_NAME = "store_attribute.v1";

/** StoreAttributeService defines the gRPC service for managing store attributes. */

export interface StoreAttributeServiceClient {
  /** rpc to get all attributes for a category */

  getAttributesByCategoryId(request: GetAttributesByCategoryIdRequest): Observable<AttributeList>;

  /** rpc to create a new attribute */

  createAttribute(request: CreateAttributeRequest): Observable<Id>;

  /** rpc to update an existing attribute */

  updateAttribute(request: UpdateAttributeRequest): Observable<Id>;

  /** rpc to delete an attribute by its ID */

  deleteAttribute(request: Id): Observable<StatusResponse>;

  /** rpc to change the position of an attribute */

  changeAttributePosition(request: ChangeAttributePositionRequest): Observable<AttributeResponse>;

  /** rpc to create or update an attribute translation */

  upsertAttributeTranslation(request: AttributeTranslationRequest): Observable<Id>;

  /** rpc to delete an attribute translation by its ID */

  deleteAttributeTranslation(request: Id): Observable<StatusResponse>;
}
// ...
export const STORE_ATTRIBUTE_SERVICE_NAME = "StoreAttributeService";